+++ src/main/java/ru/skypro/homework/service/AdsService.java
@@ -5,16 +5,15 @@ import ru.skypro.homework.dto.Ad;
 import ru.skypro.homework.dto.Ads;
 import ru.skypro.homework.dto.CreateOrUpdateAd;
 import ru.skypro.homework.dto.ExtendedAd;
-
 import java.util.List;
 
 public interface AdsService {
     Ad createAd(CreateOrUpdateAd createOrUpdateAd, MultipartFile image);
     Ads getAdAll();
     Ad updateAd(Ad ad);
-    ExtendedAd getAd(int id);
-    void deleteAd(int id);
-    List<String> updateImage(int id, MultipartFile image);
+    ExtendedAd getAd(Long id);
+    void deleteAd(Long id);
+    List<String> updateImage(Long id, MultipartFile image);
     Ads getAdsAuthorizedUser();
-    Ad updateAds(int id, CreateOrUpdateAd createOrUpdateAd);
+    Ad updateAds(Long id, CreateOrUpdateAd createOrUpdateAd);
 }
